third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/registries/element-mutation-expected.txt

This is a testharness.js-based test.
[FAIL] An element with global registry should not change its registry when run append into a shadow tree with scoped registry.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] An element with scoped registry should not change its registry when run append out of the shadow tree.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] An element with scoped registry should not change its registry when run append into another shadow tree with different scoped registry.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] A freshly created element should preserve global registry after append into and removal from a scoped shadow root.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] A freshly created element should preserve global registry when run append between scoped shadow roots.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] An element with global registry should not change its registry when run appendChild into a shadow tree with scoped registry.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] An element with scoped registry should not change its registry when run appendChild out of the shadow tree.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] An element with scoped registry should not change its registry when run appendChild into another shadow tree with different scoped registry.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] A freshly created element should preserve global registry after appendChild into and removal from a scoped shadow root.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] A freshly created element should preserve global registry when run appendChild between scoped shadow roots.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] An element with global registry should not change its registry when run prepend into a shadow tree with scoped registry.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] An element with scoped registry should not change its registry when run prepend out of the shadow tree.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] An element with scoped registry should not change its registry when run prepend into another shadow tree with different scoped registry.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] A freshly created element should preserve global registry after prepend into and removal from a scoped shadow root.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] A freshly created element should preserve global registry when run prepend between scoped shadow roots.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Declarative shadow DOM with shadowrootcustomelementregistry attribute without registry initialized should remain null registry after adoption.
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
Harness: the test ran to completion.
